+++ spec/features/transfers/users_can_edit_project_details_spec.rb
@@ -184,4 +184,39 @@ RSpec.feature "Users can edit transfer project details" do
       expect(page).to have_content("No")
     end
   end
+
+  scenario "they can change the due to an inadequate Ofsted rating response" do
+    visit project_information_path(project)
+
+    row = find("#reasonsFor .govuk-summary-list__row:nth-child(2)")
+
+    within(row) do
+      expect(page).to have_content("No")
+      click_link "Change"
+    end
+
+    within("#inadequate-ofsted") do
+      choose("Yes")
+    end
+
+    click_on "Continue"
+
+    within(row) do
+      expect(page).to have_content("Yes")
+    end
+
+    within(row) do
+      click_link "Change"
+    end
+
+    within("#inadequate-ofsted") do
+      choose("No")
+    end
+
+    click_on "Continue"
+
+    within(row) do
+      expect(page).to have_content("No")
+    end
+  end
 end