+++ CHANGELOG.md
@@ -25,6 +25,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Users can now change the advisory board date for a transfer project.
 - Users can now change the requires two improvement response for a transfer
   project.
+- Users can now change the due to inadequate Ofsted response for a transfer
+  project.
 
 ## [Release-62][release-62]